手牌数字段类型修改

+++ src/main/java/com/dx/facade/texas/resp/DxGameBillResp.java
@@ -90,5 +90,5 @@ public class DxGameBillResp implements Serializable {
     /**
      * Walks手牌数
      */
-    private Integer walksHandNumber;
+    private Long walksHandNumber;
 }
